# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Startup Hub API — idea evaluation, investor matching, and startup discovery
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type RegisterBodyRole = typeof RegisterBodyRole[keyof typeof RegisterBodyRole];


export const RegisterBodyRole = {
  founder: 'founder',
  investor: 'investor',
} as const;

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  role?: RegisterBodyRole;
  firm?: string;
  bio?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  founder: 'founder',
  investor: 'investor',
} as const;

export interface User {
  id: string;
  email: string;
  name: string;
  role?: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type IdeaStage = typeof IdeaStage[keyof typeof IdeaStage];


export const IdeaStage = {
  idea: 'idea',
  mvp: 'mvp',
  growth: 'growth',
  scale: 'scale',
} as const;

export interface Idea {
  id: string;
  userId: string;
  title: string;
  description: string;
  category: string;
  stage: IdeaStage;
  evaluationScore?: number | null;
  evaluationFeedback?: string | null;
  /** Sub-score 0-25 */
  problemClarity?: number | null;
  /** Sub-score 0-25 */
  marketSize?: number | null;
  /** Sub-score 0-25 */
  innovationScore?: number | null;
  /** Sub-score 0-25 */
  executionPotential?: number | null;
  strengths?: string[] | null;
  weaknesses?: string[] | null;
  suggestions?: string[] | null;
  marketPotential?: string | null;
  isPublic: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateIdeaBodyStage = typeof CreateIdeaBodyStage[keyof typeof CreateIdeaBodyStage];


export const CreateIdeaBodyStage = {
  idea: 'idea',
  mvp: 'mvp',
  growth: 'growth',
  scale: 'scale',
} as const;

export interface CreateIdeaBody {
  title: string;
  description: string;
  category: string;
  stage: CreateIdeaBodyStage;
  isPublic?: boolean;
}

export interface EvaluateIdeaBody {
  ideaId?: string | null;
  /** Startup name */
  name: string;
  description: string;
  /** Who the product is built for */
  target_users?: string;
  /** The core problem being solved */
  problem?: string;
  /** Optional link to a demo or prototype */
  demoLink?: string | null;
  category?: string;
  stage?: string;
}

export type EvaluationAIResultMarketPotential = typeof EvaluationAIResultMarketPotential[keyof typeof EvaluationAIResultMarketPotential];


export const EvaluationAIResultMarketPotential = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface EvaluationAIResult {
  /** Score for problem clarity (0-25) */
  problem_clarity: number;
  /** Score for market size (0-25) */
  market_size: number;
  /** Score for innovation (0-25) */
  innovation: number;
  /** Score for execution potential (0-25) */
  execution_potential: number;
  /** Total score (0-100, sum of all sub-scores) */
  total_score: number;
  strengths: string[];
  weaknesses: string[];
  /** Suggestions for improvement */
  suggestions: string[];
  market_potential: EvaluationAIResultMarketPotential;
  /** Narrative summary */
  feedback: string;
}

export interface EvaluationResult {
  /** Overall score 0-100 */
  score: number;
  marketPotential: number;
  feasibility: number;
  innovation: number;
  competition: number;
  feedback: string;
  strengths: string[];
  weaknesses: string[];
  recommendations: string[];
}

export interface Startup {
  id: string;
  title: string;
  description: string;
  category: string;
  stage: string;
  evaluationScore?: number | null;
  founderName: string;
  founderId?: string;
  createdAt: string;
}

export interface MatchInvestorsBody {
  ideaId?: string | null;
  title: string;
  description: string;
  category: string;
  stage: string;
}

export type InvestorRiskLevel = typeof InvestorRiskLevel[keyof typeof InvestorRiskLevel];


export const InvestorRiskLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface Investor {
  id: string;
  name: string;
  firm: string;
  /** Investment interest areas (e.g. AI, HealthTech) */
  interests: string[];
  riskLevel: InvestorRiskLevel;
  matchScore: number;
  /** Short explanation of why this investor matches */
  whyMatch: string;
  bio: string;
  contactHint: string;
  /** Set when this match is a registered investor user */
  userId?: string | null;
  /** True if this investor has a LiftOff account and can be messaged */
  registered?: boolean;
}

export interface ScoreBreakdown {
  problemClarity?: number | null;
  marketSize?: number | null;
  innovation?: number | null;
  executionPotential?: number | null;
}

export interface DashboardSummary {
  totalIdeas: number;
  evaluatedIdeas: number;
  averageScore?: number | null;
  topCategory?: string | null;
  investorMatches: number;
  avgBreakdown?: ScoreBreakdown;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  idea_submitted: 'idea_submitted',
  idea_evaluated: 'idea_evaluated',
  investor_matched: 'investor_matched',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  description: string;
  createdAt: string;
}

export type GetStartupsParams = {
search?: string;
category?: string;
stage?: string;
};